+++ packages/grafana-ui/src/components/DateTimePickers/DatePicker/DatePicker.tsx
@@ -2,10 +2,10 @@ import React, { memo } from 'react';
 import Calendar from 'react-calendar/dist/entry.nostyle';
 import { css } from 'emotion';
 import { GrafanaTheme2 } from '@grafana/data';
-import { useStyles2 } from '../../themes';
-import { ClickOutsideWrapper } from '../ClickOutsideWrapper/ClickOutsideWrapper';
-import { Icon } from '../Icon/Icon';
-import { getBodyStyles } from '../TimePicker/TimeRangePicker/TimePickerCalendar';
+import { useStyles2 } from '../../../themes';
+import { ClickOutsideWrapper } from '../../ClickOutsideWrapper/ClickOutsideWrapper';
+import { Icon } from '../../Icon/Icon';
+import { getBodyStyles } from '../TimeRangePicker/TimePickerCalendar';
 
 /** @public */
 export interface DatePickerProps {
