+++ packages/grafana-ui/src/components/DateTimePickers/DatePickerWithInput/DatePickerWithInput.story.tsx
@@ -2,12 +2,12 @@ import React, { useState } from 'react';
 import { DatePickerWithInput } from './DatePickerWithInput';
 
 export default {
-  title: 'Forms/DatePickerWithInput',
+  title: 'Pickers And Editors/DatePickerWithInput',
   component: DatePickerWithInput,
 };
 
 export const Basic = () => {
   const [date, setDate] = useState<Date>(new Date());
 
-  return <DatePickerWithInput width={40} value={date} onChange={(newDate) => setDate(newDate)} />;
+  return <DatePickerWithInput width={40} value={date} onChange={(newDate) => setDate(newDate as Date)} />;
 };
