+++ packages/grafana-ui/src/components/DateTimePickers/TimeZonePicker/TimeZoneOption.tsx
