+++ packages/grafana-ui/src/components/DateTimePickers/_TimeOfDayPicker.scss
