+++ packages/grafana-ui/src/components/DateTimePickers/rangeOptions.ts
