+++ packages/grafana-ui/src/components/index.scss
@@ -2,7 +2,7 @@
 @import 'Drawer/Drawer';
 @import 'RefreshPicker/RefreshPicker';
 @import 'Forms/Legacy/Select/Select';
-@import 'TimePicker/TimeOfDayPicker';
+@import 'DateTimePickers/TimeOfDayPicker';
 @import 'Tooltip/Tooltip';
 @import 'Slider/Slider';
 @import 'uPlot/Plot';
